refactor(python): split into multiple files
The one file was growing too big. Split it into multiple files for
easier maintenance.
